Moved logger implementation in new package + adjusted imports
