+++ app.py
@@ -40,6 +40,7 @@ from api.PclusterApiHandler import (
     scontrol_job,
     CLIENT_ID, CLIENT_SECRET, USER_POOL_ID
 )
+from api.logging import parse_log_entry, log_from_frontend
 from api.pcm_globals import logger
 from api.security.csrf import CSRF
 from api.security.csrf.csrf import csrf_needed
